--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,13 +1,14 @@
 package hw10programoptimization
 
 import (
-	"encoding/json"
+	"bufio"
 	"fmt"
 	"io"
-	"regexp"
+	"log"
 	"strings"
 )
 
+//go:generate easyjson -all stats.go
 type User struct {
 	ID       int
 	Name     string
@@ -31,35 +32,35 @@ func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
 type users [100_000]User
 
 func getUsers(r io.Reader) (result users, err error) {
-	content, err := io.ReadAll(r)
-	if err != nil {
-		return
-	}
+	var (
+		line []byte
+		i    int
+	)
 
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
+	scanner := bufio.NewScanner(r)
+	for scanner.Scan() {
+		line = scanner.Bytes()
 		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
+		if err = user.UnmarshalJSON(line); err != nil {
 			return
 		}
 		result[i] = user
+		i++
 	}
-	return
+
+	if err = scanner.Err(); err != nil {
+		log.Fatalf("scan file error: %v", err)
+		return
+	}
+	return result, nil
 }
 
 func countDomains(u users, domain string) (DomainStat, error) {
 	result := make(DomainStat)
 
 	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-		if err != nil {
-			return nil, err
-		}
-
-		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+		if strings.HasSuffix(user.Email, "."+domain) {
+			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]++
 		}
 	}
 	return result, nil